+++ packages/draggabletable/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 8.3.7
+
+### Patch Changes
+
+- Table Row Selection
+- Updated dependencies [undefined]
+  - @heathmont/moon-components@8.3.7
+  - @heathmont/moon-table@8.3.7
+  - @heathmont/moon-themes@8.3.7
+  - @heathmont/moon-utils@8.3.7
+
 ## 8.3.6
 
 ### Patch Changes